requirements.txt
Flask==2.3.2
openai==0.27.8
google-generativeai